+++ Source/SelfService/Web/microservice/helpers/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { capitalize, removeStringPattern, removeStringPatternAndCapitalize, getRuntimeNumberFromString } from './stringHelpers';
-export { formatTime, formatStartingDate } from './dateHelpers';
-export { DownloadLogs } from './downloadHelpers';
+export { capitalize, removeStringPattern, removeStringPatternAndCapitalize, getRuntimeNumberFromString } from './strings';
+export { formatTime, formatStartingDate } from './dates';
+export { DownloadLogs } from './downloads';
 export { formatBytes } from './numbers';
